% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pNMI}
\alias{pNMI}
\title{paralleled normalised mutual information computation}
\usage{
pNMI(X, ncore, method = "emp", type = "max.marginal", disc = "equalfreq",
  nbins = NROW(X)^(1/3), verbose = TRUE)
}
\arguments{
\item{X,}{vector/factor denoting a random variable or a data.frame denoting a
random vector where \strong{column contain variables/features and row contain outcomes/samples.}}

\item{ncore}{number of cores}

\item{method}{The name of the entropy estimator. The package implements four
estimators : "emp", "mm", "shrink", "sg" (default:"emp") - see details.}

\item{type}{method of normalization. Default is "max.marginal". Other methods
include "marginal", "joint", "min.marginal", "max.marginal", "min.conditional",
"MI", "max.conditional". See details below.}

\item{disc}{The name of the discretization method to be used :"equalfreq",
"equalwidth" or "globalequalwidth" (default : "equalfreq") - see \code{discretize}.}

\item{nbins}{Integer specifying the number of bins to be used for the
discretization. By default the number of bins is set to (N)^(1/3) where N is
the number of samples.}
}
\value{
a NMI matrix
}
\description{
Unsupervized Data Discretization and paralleled normalised mutual information computation.
}
\details{
method of the entropy estimator:
\itemize{
\item "emp" This estimator computes the entropy of the empirical probability distribution.
\item "mm" This is the Miller-Madow asymptotic bias corrected empirical estimator.
\item "shrink" This is a shrinkage estimate of the entropy of a Dirichlet probability distribution.
\item "sg" This is the Schurmann-Grassberger estimate of the entropy of a Dirichlet probability distribution.
}
type of Normalization
\itemize{
\item marginal MI = 2*( Hx + Hy - Hxy ) / ( Hx + Hy )
\item joint    MI = 2*( Hx + Hy - Hxy ) / ( Hxy )
\item min.marginal	MI = ( Hx + Hy - Hxy ) / min(Hx,Hy)
\item max.marginal MI = ( Hx + Hy - Hxy ) / max(Hx,Hy)
\item min.conditional MI = ( Hx + Hy - Hxy ) / min(Hx.y,Hy.x)
\item max.conditional MI = ( Hx + Hy - Hxy ) / max(Hx.y,Hy.x)
\item MI = no Normalization
}
}
\examples{
data(USArrests)
pNMI(USArrests, 4)
}

